Separate Modal from AddTask
- Select current project as project in AddTaskModal form
