+++ client/src/components/AddTask/AddTaskModal/styles.scss
@@ -1,4 +1,4 @@
-@import '../../stylesheets/base';
+@import '../../../stylesheets/base';
 
 .add-task-form {
     @include background-color($color-foreground);
