+++ client/src/components/AddTask/requests/TaskPostRequest.ts
@@ -15,10 +15,11 @@ import {
     alterResponseErrorToFaramError,
 } from '../../../rest';
 
-import { AddTask } from '../../AddTask';
+import { AddTaskModal } from '../../AddTask/AddTaskModal';
 
 interface Props {
-    setState: AddTask['setState'];
+    onClose(): void;
+    setState: AddTaskModal['setState'];
     setTask(value: Task): void;
 }
 
@@ -33,14 +34,7 @@ export default class TaskPostRequest implements Request < {} > {
         try {
             schema.validate(response, 'tasksPostResponse');
             this.props.setTask(response);
-            this.props.setState({
-                showModal: false,
-
-                pending: false,
-                pristine: true,
-                faramErrors: {},
-                faramValues: {},
-            });
+            this.props.onClose();
         } catch (err) {
             console.error(err);
         }
@@ -65,8 +59,15 @@ export default class TaskPostRequest implements Request < {} > {
         const request = new FgRestBuilder()
             .url(urlForTasks)
             .params(createParamsForPostTask(values))
-            .preLoad(() => { this.props.setState({ pending: true, pristine: false }); })
-            .postLoad(() => { this.props.setState({ pending: false }); })
+            .preLoad(() => {
+                this.props.setState({
+                    pending: true,
+                    pristine: false,
+                });
+            })
+            .postLoad(() => {
+                this.props.setState({ pending: false });
+            })
             .success(this.success)
             .failure(this.failure)
             .fatal(this.fatal)
