+++ client/src/views/Workspace/requests/GetTasksRequest.ts
@@ -5,7 +5,7 @@ import {
 
 import { Request } from '../../../rest/interface';
 import { Task } from '../../../redux/interface';
-import schema from '../../../schema'; 
+import schema from '../../../schema';
 import {
     urlForTasks,
     commonParamsForGet,
